Split authentication controller
- Split the authentication controller
- Moved a route definition
